+++ packages/plasma-temple/src/components/GalleryNew/SingleGallery/SingleGallery@sberbox.tsx
@@ -19,9 +19,14 @@ const StyledTitle = styled.div`
     margin-bottom: 1rem;
 `;
 
+const StyledGridWrapper = styled(CarouselGridWrapper)`
+    margin-top: -0.25rem;
+    margin-bottom: -0.25rem;
+`;
+
 const StyledCarouselItem = styled(CarouselItem)`
-    padding: 0.25rem;
-    margin-right: 0.75rem;
+    padding: 0.25rem 0;
+    margin-right: 1rem;
 `;
 
 export const SingleGallerySberBox = React.memo<SingleGalleryProps>(
@@ -82,7 +87,7 @@ export const SingleGallerySberBox = React.memo<SingleGalleryProps>(
                 className={className}
             >
                 {gallery.title && <StyledTitle>{gallery.title}</StyledTitle>}
-                <CarouselGridWrapper>
+                <StyledGridWrapper>
                     <Carousel axis="x" index={activeCard} paddingEnd="50%">
                         {gallery.items.map((item, idx) => (
                             <StyledCarouselItem scrollSnapAlign="start" key={String(item.id)}>
@@ -97,7 +102,7 @@ export const SingleGallerySberBox = React.memo<SingleGalleryProps>(
                             </StyledCarouselItem>
                         ))}
                     </Carousel>
-                </CarouselGridWrapper>
+                </StyledGridWrapper>
             </StyledContainer>
         );
     },