+++ packages/plasma-temple/src/components/Header/Header.tsx
@@ -6,6 +6,7 @@ import { useNewHeaderProps, useHeaderProps } from './useHeaderProps';
 
 export { useHeaderProps };
 
+/** Заголовок Canvas App. Работает только с компонентом PlasmaApp */
 export const Header: React.FC<HeaderProps> = (props) => {
     const headerProps = useNewHeaderProps(props);