+++ CHANGELOG.md
@@ -1,3 +1,95 @@
+# (Thu Jun 09 2022)
+
+#### 🚀 Enhancement
+
+- `@salutejs/plasma-icons@1.86.0`
+  - feat(plasma-icons): Add Sber Icons [#63](https://github.com/salute-developers/plasma/pull/63) ([@andart71](https://github.com/andart71))
+
+#### Authors: 1
+
+- Andrey Loshkarev ([@andart71](https://github.com/andart71))
+
+---
+
+# (Thu Jun 09 2022)
+
+#### 🐛 Bug Fix
+
+- `@salutejs/plasma-ui@1.108.2`
+  - fix(plasma-ui): Add min-height in StyledMediaSlot for `ProductCard` component [#61](https://github.com/salute-developers/plasma/pull/61) ([@neretin-trike](https://github.com/neretin-trike))
+- `@salutejs/plasma-web@1.107.1`
+  - fix(plasma-web): Remove children from DropdownProps [#60](https://github.com/salute-developers/plasma/pull/60) ([@andart71](https://github.com/andart71))
+
+#### Authors: 2
+
+- Andrey Loshkarev ([@andart71](https://github.com/andart71))
+- neretinaa ([@neretin-trike](https://github.com/neretin-trike))
+
+---
+
+# (Thu Jun 09 2022)
+
+#### 🐛 Bug Fix
+
+- `@salutejs/plasma-temple@1.72.2`
+  - fix(plasma-temple): useSpatNavBetweenTargets - fix ssr render [#59](https://github.com/salute-developers/plasma/pull/59) ([@kvpolyanskiy](https://github.com/kvpolyanskiy))
+
+#### Authors: 1
+
+- [@kvpolyanskiy](https://github.com/kvpolyanskiy)
+
+---
+
+# (Tue Jun 07 2022)
+
+#### 🐛 Bug Fix
+
+- `@salutejs/plasma-b2c@1.72.1`, `@salutejs/plasma-ui@1.108.1`
+  - fix(plasma-b2c, plasma-ui): fix scroll handle position passed to callbacks [#48](https://github.com/salute-developers/plasma/pull/48) ([@Gerpea](https://github.com/Gerpea))
+
+#### Authors: 1
+
+- German Cyganov ([@Gerpea](https://github.com/Gerpea))
+
+---
+
+# (Tue Jun 07 2022)
+
+#### 🚀 Enhancement
+
+- `@salutejs/plasma-temple@1.72.0`
+  - feat(plasma-temple): fix gallery offset between title and gallery [#42](https://github.com/salute-developers/plasma/pull/42) ([@kvpolyanskiy](https://github.com/kvpolyanskiy))
+  - feat(plasma-temple): fix gallery paddings [#42](https://github.com/salute-developers/plasma/pull/42) ([@kvpolyanskiy](https://github.com/kvpolyanskiy))
+  - feat(plasma-temple): gallery skeleton [#42](https://github.com/salute-developers/plasma/pull/42) ([@kvpolyanskiy](https://github.com/kvpolyanskiy))
+
+#### 🐛 Bug Fix
+
+- `@salutejs/plasma-core@1.62.0`, `@salutejs/plasma-web@1.107.0`
+  - test(plasma-b2c,plasma-web,plasma-ui): Add test and snapshot for `TextField` component [#47](https://github.com/salute-developers/plasma/pull/47) ([@neretin-trike](https://github.com/neretin-trike))
+- `@salutejs/plasma-b2c@1.72.0`, `@salutejs/plasma-ui@1.108.0`, `@salutejs/plasma-web@1.107.0`
+  - fix(plasma-web,plasma-b2c,plasma-ui): Add onChange callback for maxLength check in `TextField` [#47](https://github.com/salute-developers/plasma/pull/47) ([@neretin-trike](https://github.com/neretin-trike))
+
+#### Authors: 2
+
+- [@kvpolyanskiy](https://github.com/kvpolyanskiy)
+- neretinaa ([@neretin-trike](https://github.com/neretin-trike))
+
+---
+
+# (Tue Jun 07 2022)
+
+#### 🚀 Enhancement
+
+- `@salutejs/plasma-temple@1.71.0`
+  - feat(plasma-temple): fixes [#52](https://github.com/salute-developers/plasma/pull/52) ([@kvpolyanskiy](https://github.com/kvpolyanskiy))
+  - feat(plasma-temple): Use ScreenProvider and AssistantProvider under the hood of PlasmaApp [#52](https://github.com/salute-developers/plasma/pull/52) ([@kvpolyanskiy](https://github.com/kvpolyanskiy))
+
+#### Authors: 1
+
+- [@kvpolyanskiy](https://github.com/kvpolyanskiy)
+
+---
+
 # (Thu Jun 02 2022)
 
 #### 🐛 Bug Fix